[deploy] explicitly error if artifact URLs are blank

+++ services/pelias/init_elastic.sh
@@ -9,11 +9,14 @@ elif [ -f "${ELASTICSEARCH_ARTIFACT_SOURCE_PATH}" ]; then
     echo "Extracting artifact."
     mkdir -p /usr/share/elasticsearch/data
     tar -xJf "${ELASTICSEARCH_ARTIFACT_SOURCE_PATH}" -C /usr/share/elasticsearch/data
-else
+elif [ ! -z "${ELASTICSEARCH_ARTIFACT_URL}" ]; then
     echo "Downloading and extracting artifact."
     rm -fr /tmp/elasticsearch.download
     mkdir -p /tmp/elasticsearch.download
     wget --tries=100 -O- "${ELASTICSEARCH_ARTIFACT_URL}" | tar -xJ -C /tmp/elasticsearch.download
     mv /tmp/elasticsearch.download/nodes /usr/share/elasticsearch/data/nodes
     rmdir /tmp/elasticsearch.download
+else
+    echo "No elasticsearch artifact available."
+    exit 1
 fi
