+++ services/pelias/init_placeholder.sh
@@ -9,10 +9,13 @@ elif [ -f "${PLACEHOLDER_ARTIFACT_SOURCE_PATH}" ]; then
     echo "Extracting artifact."
     mkdir -p /data/placeholder
     tar -xJf "${PLACEHOLDER_ARTIFACT_SOURCE_PATH}" -C /data/placeholder
-else
+elif [ ! -z "${PLACEHOLDER_ARTIFACT_URL}" ]; then
     echo "Downloading and extracting artifact."
     rm -fr /tmp/placeholder.download
     mkdir -p /tmp/placeholder.download
     wget --tries=100 -O- "${PLACEHOLDER_ARTIFACT_URL}" | tar -xJ -C /tmp/placeholder.download
     mv /tmp/placeholder.download /data/placeholder
+else
+    echo "No placeholder artifact available."
+    exit 1
 fi
